requirements.txt
annotated-types==0.7.0
greenlet==3.2.4
psycopg2==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.14.1
black<25.0
click==8.2.1
flake8<7.0
mccabe==0.7.0
mypy<1.12
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.11.1
pyflakes==3.1.0
typing_extensions==4.14.1
